moved templateCompilerProvider to strings
